[chore] Move batcher to the new QueueBatcher package
