+++ exporter/exporterhelper/internal/queuebatch/batcher.go
@@ -1,14 +1,13 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-package batcher // import "go.opentelemetry.io/collector/exporter/exporterhelper/internal/batcher"
+package queuebatch // import "go.opentelemetry.io/collector/exporter/exporterhelper/internal/queuebatch"
 
 import (
 	"context"
 
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/exporter/exporterbatcher"
-	"go.opentelemetry.io/collector/exporter/exporterhelper/internal/queuebatch"
 	"go.opentelemetry.io/collector/exporter/exporterhelper/internal/request"
 	"go.opentelemetry.io/collector/exporter/exporterhelper/internal/sender"
 )
@@ -16,7 +15,7 @@ import (
 // Batcher is in charge of reading items from the queue and send them out asynchronously.
 type Batcher[K any] interface {
 	component.Component
-	Consume(context.Context, K, queuebatch.Done)
+	Consume(context.Context, K, Done)
 }
 
 func NewBatcher(batchCfg exporterbatcher.Config, exportFunc sender.SendFunc[request.Request], maxWorkers int) (Batcher[request.Request], error) {
