+++ exporter/exporterhelper/internal/queuebatch/default_batcher_test.go
@@ -1,7 +1,7 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-package batcher
+package queuebatch
 
 import (
 	"context"
