+++ exporter/exporterhelper/internal/queuebatch/disabled_batcher.go
@@ -1,13 +1,12 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-package batcher // import "go.opentelemetry.io/collector/exporter/exporterhelper/internal/batcher"
+package queuebatch // import "go.opentelemetry.io/collector/exporter/exporterhelper/internal/queuebatch"
 
 import (
 	"context"
 
 	"go.opentelemetry.io/collector/component"
-	"go.opentelemetry.io/collector/exporter/exporterhelper/internal/queuebatch"
 	"go.opentelemetry.io/collector/exporter/exporterhelper/internal/sender"
 )
 
@@ -19,7 +18,7 @@ type disabledBatcher[T any] struct {
 	consumeFunc sender.SendFunc[T]
 }
 
-func (db *disabledBatcher[T]) Consume(ctx context.Context, req T, done queuebatch.Done) {
+func (db *disabledBatcher[T]) Consume(ctx context.Context, req T, done Done) {
 	done.OnDone(db.consumeFunc(ctx, req))
 }
